refactor: extract protocol constants to src/protocol/constants.ts
