+++ src/config.ts
@@ -1,11 +1,1 @@
-export const DEFAULT_CONFIG = {
-	// Multicast (UDP)
-	MULTICAST_PORT: 53317,
-	MULTICAST_ADDRESS: "224.0.0.167",
-
-	// HTTP (TCP)
-	HTTP_PORT: 53317,
-
-	// Protocol version
-	PROTOCOL_VERSION: "2.1"
-}
+export * from "./protocol/constants.ts"
